feat: add endpoints for myflight

+++ src/main/java/com/shane/iredeem/controller/ItineraryController.java
@@ -28,7 +28,7 @@ public class ItineraryController {
     }
 
     @GetMapping("/itineraries/{id}")
-    public Itinerary getItinerary(@PathVariable("id") Integer id) {
+    public Itinerary getItinerary(@PathVariable("id") Long id) {
         return itineraryService.getItinerary(id).orElse(null);
     }
